code/requirement.txt
opencv-python
pillow>=8.0
pandas
numpy
scipy
scikit-image>=0.17.2
scikit-learn>=0.23.0
matplotlib>=3.3.4
ipython
jupyter
monai==0.4.0
torch==1.7.1
torchvision==0.5.0
tensorboard
captum
seaborn
psutil
tqdm
nibabel